move game logistic to App.jsx
